requirements.txt
Pillow==9.5.0
torch==2.0.1
diffusers==0.18.2
langchain-groq==0.0.4
python-dotenv==1.0.0
gradio==3.42.0
